Dinamic galery and code refactoring
