+++ src/Styles/body.css
@@ -20,7 +20,7 @@
 }
 
 .news{
-    margin-bottom: 2%;
+    margin-bottom: 3%;
 }
 
 .contact{
