--- src/components/Body/Galery/Accessories/Accessories.js
@@ -1,7 +1,0 @@
-const Accessories = () => {
-    return(
-        <h2>Accessories</h2>
-    );
-}
-
-export default Accessories;
