--- src/components/Body/Galery/Accessories/index.js
@@ -1,1 +1,0 @@
-export { default } from './Accessories';
