--- src/components/Body/Galery/Bows/index.js
@@ -1,1 +1,0 @@
-export { default } from './Bows';
