--- src/components/Body/Galery/Dresses/Dresses.js
@@ -1,7 +1,0 @@
-const Dresses = () => {
-    return (
-        <h2>Dresses</h2>
-    );
-};
-
-export default Dresses;
