--- src/components/Body/Galery/Dresses/index.js
@@ -1,1 +1,0 @@
-export { default } from './Dresses';
